Update 23/8/21
Fixed mobile menu and scroll effects problems with the page.

Problem:
Each time the hambugger menu was clicked, the mobile drop down menu appeared and the hambugger menu crossed and became an X. But when you clicked the X to close the drop down menu, the page would scroll back to the top.

Solution:
Added an event listener to the div containing the mobile menu, and closed it on click. This prevents the need to close the menu in the middle of the page - which ordinarily would cause it to scroll back to the top.